+++ ReactAndroid/src/main/java/com/facebook/react/uimanager/UIImplementation.java
@@ -124,22 +124,6 @@ public class UIImplementation {
     }
   }
 
-  public void dropViews(ReadableArray viewTags) {
-    int size = viewTags.size(), realViewsCount = 0;
-    int realViewTags[] = new int[size];
-    for (int i = 0; i < size; i++) {
-      int tag = viewTags.getInt(i);
-      ReactShadowNode cssNode = mShadowNodeRegistry.getNode(tag);
-      if (!cssNode.isVirtual()) {
-        realViewTags[realViewsCount++] = tag;
-      }
-      mShadowNodeRegistry.removeNode(tag);
-    }
-    if (realViewsCount > 0) {
-      mNativeViewHierarchyOptimizer.handleDropViews(realViewTags, realViewsCount);
-    }
-  }
-
   /**
    * Invoked by React to create a new node with a given tag has its properties changed.
    */
@@ -509,6 +493,7 @@ public class UIImplementation {
 
   private void removeShadowNode(ReactShadowNode nodeToRemove) {
     mNativeViewHierarchyOptimizer.handleRemoveNode(nodeToRemove);
+    mShadowNodeRegistry.removeNode(nodeToRemove.getReactTag());
     for (int i = nodeToRemove.getChildCount() - 1; i >= 0; i--) {
       removeShadowNode(nodeToRemove.getChildAt(i));
     }
