+++ ReactAndroid/src/main/java/com/facebook/react/uimanager/UIManagerModuleConstantsHelper.java
@@ -25,7 +25,6 @@ import com.facebook.react.common.MapBuilder;
 
   private static final String CUSTOM_BUBBLING_EVENT_TYPES_KEY = "customBubblingEventTypes";
   private static final String CUSTOM_DIRECT_EVENT_TYPES_KEY = "customDirectEventTypes";
-  private static final String LAYOUT_ONLY_PROPS = "layoutOnlyProps";
 
   /**
    * Generates map of constants that is then exposed by {@link UIManagerModule}. The constants map
@@ -76,7 +75,6 @@ import com.facebook.react.common.MapBuilder;
 
     constants.put(CUSTOM_BUBBLING_EVENT_TYPES_KEY, bubblingEventTypesConstants);
     constants.put(CUSTOM_DIRECT_EVENT_TYPES_KEY, directEventTypesConstants);
-    constants.put(LAYOUT_ONLY_PROPS, UIManagerModuleConstants.getLayoutOnlyPropsConstants());
 
     return constants;
   }
